Rename 'components' folder to 'ui'
