+++ src/constants/globalConfig.ts
@@ -1,24 +1,24 @@
+import { Dream } from "../models/types.js";
+
 export const INPUT_MAX_LENGTH = 50;
 
-// import { Dream } from "./types";
-
-// export const dreams: Dream[] = [
-//   {
-//     id: 1,
-//     name: "Lära mig HTML/CSS",
-//     theme: "teknikdrömmar",
-//     checked: true,
-//   },
-//   {
-//     id: 2,
-//     name: "Lära mig TypeScript",
-//     theme: "teknikdrömmar",
-//     checked: false,
-//   },
-//   {
-//     id: 3,
-//     name: "En dröm som tar flera rader lorem ipsum",
-//     theme: "vardagsdrömmar",
-//     checked: false,
-//   },
-// ];
+export const dreams: Dream[] = [
+  {
+    id: 1,
+    name: "Lära mig HTML/CSS",
+    theme: "teknikdrömmar",
+    isChecked: true,
+  },
+  {
+    id: 2,
+    name: "Lära mig TypeScript",
+    theme: "teknikdrömmar",
+    isChecked: false,
+  },
+  {
+    id: 3,
+    name: "En dröm som tar flera rader lorem ipsum",
+    theme: "vardagsdrömmar",
+    isChecked: false,
+  },
+];
